added common args file and moved files around
