+++ experiments/kitchen/plan2explore/plan2explore_intrinsic_extrinsic_goals.py
@@ -1,4 +1,3 @@
-import argparse
 import random
 import subprocess
