+++ experiments/ll_raps/dreamer_v2_single_task_low_level_primitives.py
@@ -1,45 +1,40 @@
-import argparse
 import random
 import subprocess
 
 import rlkit.util.hyperparameter as hyp
 from rlkit.launchers.launcher_util import run_experiment
+from rlkit.torch.model_based.dreamer.experiments.arguments import get_args
 from rlkit.torch.model_based.dreamer.experiments.experiment_utils import (
     preprocess_variant,
 )
 from rlkit.torch.model_based.dreamer.experiments.ll_raps_experiment import experiment
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--exp_prefix", type=str, default="test")
-    parser.add_argument("--num_seeds", type=int, default=1)
-    parser.add_argument("--mode", type=str, default="local")
-    parser.add_argument("--debug", action="store_true", default=False)
-    args = parser.parse_args()
+    args = get_args()
     if args.debug:
         algorithm_kwargs = dict(
             num_epochs=5,
-            num_eval_steps_per_epoch=10,
-            num_expl_steps_per_train_loop=50,
-            min_num_steps_before_training=10,
+            num_eval_steps_per_epoch=5,
+            num_expl_steps_per_train_loop=5,
+            min_num_steps_before_training=0,
             num_pretrain_steps=10,
             num_train_loops_per_epoch=1,
-            num_trains_per_train_loop=10,
-            batch_size=200,
+            num_trains_per_train_loop=1,
+            batch_size=25,
             max_path_length=5,
         )
         exp_prefix = "test" + args.exp_prefix
     else:
         algorithm_kwargs = dict(
-            num_epochs=250,
+            num_epochs=1000,
             num_eval_steps_per_epoch=30,
             min_num_steps_before_training=2500,
-            num_pretrain_steps=100,
+            num_pretrain_steps=1000,
             max_path_length=5,
-            batch_size=200,
-            num_expl_steps_per_train_loop=30 * 2,  # 5*(5+1) one trajectory per vec env
-            num_train_loops_per_epoch=40 // 2,  # 1000//(5*5)
-            num_trains_per_train_loop=10 * 2,  # 400//40
+            batch_size=100,
+            num_expl_steps_per_train_loop=30,
+            num_train_loops_per_epoch=10,
+            num_trains_per_train_loop=100,
         )
         exp_prefix = args.exp_prefix
     variant = dict(
@@ -93,10 +88,10 @@ if __name__ == "__main__":
             pred_discount_num_layers=3,
             gru_layer_norm=True,
             std_act="sigmoid2",
-            depth=32,
+            use_prior_instead_of_posterior=True,
         ),
         trainer_kwargs=dict(
-            adam_eps=1e-5,
+            adam_eps=1e-8,
             discount=0.8,
             lam=0.95,
             forward_kl=False,
@@ -106,7 +101,7 @@ if __name__ == "__main__":
             transition_loss_scale=0.8,
             actor_lr=8e-5,
             vf_lr=8e-5,
-            world_model_lr=3e-4,
+            world_model_lr=1e-3,
             reward_loss_scale=2.0,
             use_pred_discount=True,
             policy_gradient_loss_scale=1.0,
@@ -114,7 +109,7 @@ if __name__ == "__main__":
             target_update_period=100,
             detach_rewards=False,
             imagination_horizon=5,
-            wm_loss_scale=-1,
+            weight_decay=0.0,
         ),
         num_expl_envs=5,
         num_eval_envs=1,
@@ -124,7 +119,8 @@ if __name__ == "__main__":
         mlp_hidden_sizes=[512, 512],
         prioritize_fraction=0.0,
         uniform_priorities=True,
-        num_low_level_actions_per_primitive=10,
+        # unsubsampled_rollout=True,
+        # generate_video=True,
     )
 
     search_space = {
@@ -133,15 +129,38 @@ if __name__ == "__main__":
             "disassemble-v2",
             "sweep-into-v2",
             "soccer-v2",
+            # "drawer-close-v2",
         ],
-        "model_kwargs.use_prior_instead_of_posterior": [True, False],
-        "num_low_level_actions_per_primitive": [10, 5],
+        "algorithm_kwargs.num_train_loops_per_epoch": [10],
+        "algorithm_kwargs.num_expl_steps_per_train_loop": [30],
+        "algorithm_kwargs.num_pretrain_steps": [1000],
+        "algorithm_kwargs.num_trains_per_train_loop": [100],
+        "algorithm_kwargs.min_num_steps_before_training": [2500],
+        "algorithm_kwargs.batch_size": [100],
+        "num_low_level_actions_per_primitive": [10],
+        "trainer_kwargs.batch_length": [50],
+        # "replay_buffer_path": [
+        #     "/home/mdalal/research/skill_learn/hrl-exp/data/world_model_data/assembly_demo_data.hdf5"
+        # ],
+        # "trainer_kwargs.binarize_rewards": [True, False],
+        # "model_kwargs.reward_classifier": [True, False ],
+        # "primitive_embedding": [True, False],
+        # "prioritize_fraction": [0.25],
+        # "uniform_priorities": [False],
+        # "models_path": [
+        # "/home/mdalal/research/skill_learn/rlkit/data/01-17-ll-raps-mw-no-transposes/01-17-ll_raps_mw_no_transposes_2022_01_17_15_49_24_0000--s-61010/"
+        # "/home/mdalal/research/skill_learn/rlkit/data/01-18-ll-raps-mw-no-transposes/01-18-ll_raps_mw_no_transposes_2022_01_18_00_41_14_0000--s-64192/"
+        # ],
     }
     sweeper = hyp.DeterministicHyperparameterSweeper(
         search_space,
         default_parameters=variant,
     )
     for exp_id, variant in enumerate(sweeper.iterate_hyperparameters()):
+        if args.debug:
+            variant["algorithm_kwargs"]["num_pretrain_steps"] = 1
+            variant["algorithm_kwargs"]["min_num_steps_before_training"] = 10
+            variant["algorithm_kwargs"]["num_trains_per_train_loop"] = 1
         variant["replay_buffer_size"] = int(
             3e6 / (variant["num_low_level_actions_per_primitive"] * 5 + 1)
         )
@@ -151,22 +170,15 @@ if __name__ == "__main__":
         variant["env_kwargs"]["num_low_level_actions_per_primitive"] = variant[
             "num_low_level_actions_per_primitive"
         ]
-        # variant[
-        #     "eval_buffer_path"
-        # ] = "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_{}_T_{}_E_{}_P_{}_raps_ll_hl_even_rt_{}.hdf5".format(
-        #     5,
-        #     100,
-        #     10,
-        #     variant["num_low_level_actions_per_primitive"],
-        #     variant["env_name"],
-        # )
-        variant["trainer_kwargs"]["num_world_model_training_iterations"] = (
-            400 // variant["algorithm_kwargs"]["batch_size"]
+        variant[
+            "eval_buffer_path"
+        ] = "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_{}_T_{}_E_{}_P_{}_raps_ll_hl_even_rt_{}.hdf5".format(
+            5,
+            100,
+            10,
+            variant["num_low_level_actions_per_primitive"],
+            variant["env_name"],
         )
-        if variant["trainer_kwargs"]["wm_loss_scale"] == -1:
-            variant["trainer_kwargs"]["wm_loss_scale"] = 1 / (
-                variant["trainer_kwargs"]["num_world_model_training_iterations"]
-            )
         variant = preprocess_variant(variant, args.debug)
         for _ in range(args.num_seeds):
             seed = random.randint(0, 100000)
