+++ experiments/ll_raps/dreamer_v2_single_task_low_level_primitives_raps_params.py
@@ -1,62 +1,61 @@
-import argparse
 import random
 import subprocess
 
 import rlkit.util.hyperparameter as hyp
 from rlkit.launchers.launcher_util import run_experiment
+from rlkit.torch.model_based.dreamer.experiments.arguments import get_args
 from rlkit.torch.model_based.dreamer.experiments.experiment_utils import (
     preprocess_variant,
 )
 from rlkit.torch.model_based.dreamer.experiments.ll_raps_experiment import experiment
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--exp_prefix", type=str, default="test")
-    parser.add_argument("--num_seeds", type=int, default=1)
-    parser.add_argument("--mode", type=str, default="local")
-    parser.add_argument("--debug", action="store_true", default=False)
-    args = parser.parse_args()
+    args = get_args()
     if args.debug:
         algorithm_kwargs = dict(
             num_epochs=5,
-            num_eval_steps_per_epoch=5,
-            num_expl_steps_per_train_loop=5,
-            min_num_steps_before_training=0,
+            num_eval_steps_per_epoch=10,
+            num_expl_steps_per_train_loop=50,
+            min_num_steps_before_training=10,
             num_pretrain_steps=10,
             num_train_loops_per_epoch=1,
-            num_trains_per_train_loop=1,
-            batch_size=25,
+            num_trains_per_train_loop=10,
+            batch_size=200,
             max_path_length=5,
         )
         exp_prefix = "test" + args.exp_prefix
     else:
         algorithm_kwargs = dict(
-            num_epochs=1000,
+            num_epochs=250,
             num_eval_steps_per_epoch=30,
             min_num_steps_before_training=2500,
-            num_pretrain_steps=1000,
+            num_pretrain_steps=100,
             max_path_length=5,
-            batch_size=100,
-            num_expl_steps_per_train_loop=30,
-            num_train_loops_per_epoch=10,
-            num_trains_per_train_loop=100,
+            batch_size=200,
+            num_expl_steps_per_train_loop=30 * 2,  # 5*(5+1) one trajectory per vec env
+            num_train_loops_per_epoch=40 // 2,  # 1000//(5*5)
+            num_trains_per_train_loop=10 * 2,  # 400//40
         )
         exp_prefix = args.exp_prefix
     variant = dict(
-        algorithm="DreamerV2",
+        algorithm="LLRAPS",
         version="normal",
         replay_buffer_size=int(1.2e4),
         algorithm_kwargs=algorithm_kwargs,
         use_raw_actions=False,
-        env_suite="kitchen",
+        env_suite="metaworld",
         pass_render_kwargs=True,
         env_kwargs=dict(
-            dense=False,
-            image_obs=True,
-            action_scale=1.4,
-            use_workspace_limits=True,
             control_mode="primitives",
-            num_low_level_actions_per_primitive=10,
+            action_scale=1,
+            max_path_length=5,
+            reward_type="sparse",
+            camera_settings={
+                "distance": 0.38227044687537043,
+                "lookat": [0.21052547, 0.32329237, 0.587819],
+                "azimuth": 141.328125,
+                "elevation": -53.203125160653144,
+            },
             usage_kwargs=dict(
                 use_dm_backend=True,
                 use_raw_action_wrappers=False,
@@ -64,9 +63,10 @@ if __name__ == "__main__":
                 max_path_length=5,
                 unflatten_images=False,
             ),
-            image_kwargs=dict(),
+            image_kwargs=dict(imwidth=64, imheight=64),
             collect_primitives_info=True,
             render_intermediate_obs_to_info=True,
+            num_low_level_actions_per_primitive=10,
         ),
         actor_kwargs=dict(
             discrete_continuous_dist=True,
@@ -88,10 +88,10 @@ if __name__ == "__main__":
             pred_discount_num_layers=3,
             gru_layer_norm=True,
             std_act="sigmoid2",
-            use_prior_instead_of_posterior=True,
+            depth=32,
         ),
         trainer_kwargs=dict(
-            adam_eps=1e-8,
+            adam_eps=1e-5,
             discount=0.8,
             lam=0.95,
             forward_kl=False,
@@ -101,7 +101,7 @@ if __name__ == "__main__":
             transition_loss_scale=0.8,
             actor_lr=8e-5,
             vf_lr=8e-5,
-            world_model_lr=1e-3,
+            world_model_lr=3e-4,
             reward_loss_scale=2.0,
             use_pred_discount=True,
             policy_gradient_loss_scale=1.0,
@@ -109,8 +109,7 @@ if __name__ == "__main__":
             target_update_period=100,
             detach_rewards=False,
             imagination_horizon=5,
-            batch_length=100,
-            weight_decay=0.0,
+            wm_loss_scale=-1,
         ),
         num_expl_envs=5,
         num_eval_envs=1,
@@ -119,44 +118,50 @@ if __name__ == "__main__":
         low_level_action_dim=9,
         mlp_hidden_sizes=[512, 512],
         prioritize_fraction=0.0,
+        uniform_priorities=True,
+        num_low_level_actions_per_primitive=10,
     )
 
     search_space = {
         "env_name": [
-            "microwave",
-            "kettle",
-            "slide_cabinet",
-            "top_left_burner",
-            "hinge_cabinet",
-            "light_switch",
+            "assembly-v2",
+            "disassemble-v2",
+            "sweep-into-v2",
+            "soccer-v2",
         ],
-        "algorithm_kwargs.num_train_loops_per_epoch": [10],
-        "algorithm_kwargs.num_expl_steps_per_train_loop": [30],
-        "algorithm_kwargs.num_pretrain_steps": [1000],
-        "algorithm_kwargs.num_trains_per_train_loop": [100],
-        "algorithm_kwargs.min_num_steps_before_training": [2500],
-        "algorithm_kwargs.batch_size": [200],
-        "num_low_level_actions_per_primitive": [5],
-        "trainer_kwargs.batch_length": [50],
+        "model_kwargs.use_prior_instead_of_posterior": [True, False],
+        "num_low_level_actions_per_primitive": [10, 5],
     }
     sweeper = hyp.DeterministicHyperparameterSweeper(
         search_space,
         default_parameters=variant,
     )
     for exp_id, variant in enumerate(sweeper.iterate_hyperparameters()):
-        if args.debug:
-            variant["algorithm_kwargs"]["num_pretrain_steps"] = 1
-            variant["algorithm_kwargs"]["min_num_steps_before_training"] = 10
-            variant["algorithm_kwargs"]["num_trains_per_train_loop"] = 1
         variant["replay_buffer_size"] = int(
             3e6 / (variant["num_low_level_actions_per_primitive"] * 5 + 1)
         )
         variant["trainer_kwargs"]["batch_length"] = int(
-            variant["num_low_level_actions_per_primitive"] * 5
+            variant["num_low_level_actions_per_primitive"] * 5 + 1
         )
         variant["env_kwargs"]["num_low_level_actions_per_primitive"] = variant[
             "num_low_level_actions_per_primitive"
         ]
+        # variant[
+        #     "eval_buffer_path"
+        # ] = "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_{}_T_{}_E_{}_P_{}_raps_ll_hl_even_rt_{}.hdf5".format(
+        #     5,
+        #     100,
+        #     10,
+        #     variant["num_low_level_actions_per_primitive"],
+        #     variant["env_name"],
+        # )
+        variant["trainer_kwargs"]["num_world_model_training_iterations"] = (
+            400 // variant["algorithm_kwargs"]["batch_size"]
+        )
+        if variant["trainer_kwargs"]["wm_loss_scale"] == -1:
+            variant["trainer_kwargs"]["wm_loss_scale"] = 1 / (
+                variant["trainer_kwargs"]["num_world_model_training_iterations"]
+            )
         variant = preprocess_variant(variant, args.debug)
         for _ in range(args.num_seeds):
             seed = random.randint(0, 100000)
