+++ experiments/ll_raps/train_primitives_with_frozen_world_model.py
@@ -4,17 +4,13 @@ import subprocess
 
 import rlkit.util.hyperparameter as hyp
 from rlkit.launchers.launcher_util import run_experiment
+from rlkit.torch.model_based.dreamer.experiments.arguments import get_args
 from rlkit.torch.model_based.dreamer.experiments.world_model_training_experiment import (
     experiment,
 )
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--exp_prefix", type=str, default="test")
-    parser.add_argument("--num_seeds", type=int, default=1)
-    parser.add_argument("--mode", type=str, default="local")
-    parser.add_argument("--debug", action="store_true", default=False)
-    args = parser.parse_args()
+    args = get_args()
     if args.debug:
         algorithm_kwargs = dict()
         exp_prefix = "test" + args.exp_prefix
@@ -22,7 +18,6 @@ if __name__ == "__main__":
         algorithm_kwargs = dict()
         exp_prefix = args.exp_prefix
     variant = dict(
-        plotting_period=25,
         low_level_primitives=True,
         num_low_level_actions_per_primitive=100,
         low_level_action_dim=9,
@@ -71,32 +66,31 @@ if __name__ == "__main__":
             use_prior_instead_of_posterior=True,
         ),
         optimizer_kwargs=dict(
-            lr=3e-4,
-            eps=1e-5,
-            weight_decay=0.0,
+            lr=1e-3,
         ),
-        gradient_clip=100,
         dataloader_kwargs=dict(
-            batch_len=100,
-            batch_size=25,
+            batch_len=50,
+            batch_size=50,
             train_test_split=0.8,
             randomize_batch_len=True,
         ),
-        num_epochs=1000,
+        mlp_hidden_sizes=(512, 512),
+        gradient_clip=0,
         world_model_path="/home/mdalal/research/skill_learn/rlkit/data/11-19-train-wm-even-100-1/11-19-train_wm_even_100_1_2021_11_19_21_48_13_0000--s-23958/models/world_model.pt",
-        train_test_split=0.8,
         clone_primitives=False,
-        clone_primitives_separately=False,
+        clone_primitives_separately=True,
+        plotting_period=100,
+        num_epochs=10000,
+        visualize_wm_from_path=False,
     )
 
     search_space = {
-        "num_epochs": [10000],
-        "visualize_wm_from_path": [False],
-        "dataloader_kwargs.randomize_batch_len": [False, True],
+        "clone_primitives": [True],
+        "dataloader_kwargs.batch_len": [100],
+        "clone_primitives_separately": [False],
+        "mlp_hidden_sizes": [(512, 512)],
         "datafile": [
-            "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_drawer-close-v2.hdf5",
-            "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_soccer-v2.hdf5",
-            "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_reach-v2.hdf5",
+            "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even.hdf5",
         ],
     }
     sweeper = hyp.DeterministicHyperparameterSweeper(
@@ -104,26 +98,15 @@ if __name__ == "__main__":
         default_parameters=variant,
     )
     for exp_id, variant in enumerate(sweeper.iterate_hyperparameters()):
-        if variant["dataloader_kwargs"]["randomize_batch_len"]:
-            variant["plotting_period"] = 100
-        else:
-            variant["plotting_period"] = 1
-        if (
-            variant["datafile"]
-            == "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_drawer-close-v2.hdf5"
-        ):
-            variant["env_name"] = "drawer-close-v2"
-        elif (
-            variant["datafile"]
-            == "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_soccer-v2.hdf5"
-        ):
-            variant["env_name"] = "soccer-v2"
-        elif (
-            variant["datafile"]
-            == "/home/mdalal/research/skill_learn/rlkit/data/world_model_data/wm_H_5_T_25_E_50_P_100_raps_ll_hl_even_rt_reach-v2.hdf5"
-        ):
-            variant["env_name"] = "reach-v2"
         for _ in range(args.num_seeds):
+            if not variant["dataloader_kwargs"]["randomize_batch_len"]:
+                variant["plotting_period"] = 1
+            if variant["clone_primitives"] and variant["clone_primitives_separately"]:
+                continue
+            if not (
+                variant["clone_primitives"] or variant["clone_primitives_separately"]
+            ):
+                continue
             seed = random.randint(0, 100000)
             variant["seed"] = seed
             variant["exp_id"] = exp_id
