+++ experiments/robosuite/dreamer/dreamer_v2_single_task_primitives_lift.py
@@ -1,4 +1,3 @@
-import argparse
 import random
 import subprocess
