+++ experiments/robosuite/ppo/ppo_single_task_primitives_state.py
@@ -1,4 +1,3 @@
-import argparse
 import os
 import random
 import subprocess
