+++ rlkit/torch/model_based/dreamer/experiments/arguments.py
@@ -1,0 +1,16 @@
+import argparse
+
+
+def get_args():
+    parser = argparse.ArgumentParser(description="Experiment Launcher Arguments")
+
+    parser.add_argument("--exp_prefix", type=str, default="test")
+    parser.add_argument("--num_seeds", type=int, default=1)
+    parser.add_argument("--mode", type=str, default="local")
+    parser.add_argument("--debug", action="store_true", default=False)
+
+    # parse arguments
+    args = parser.parse_args()
+    args.exp_prefix = "test" + args.exp_prefix
+
+    return args
